requirements.txt
xlwt==1.3.0
pillow==10.3.0
